navbar

+++ client/src/services/auth/authService.ts
@@ -10,53 +10,30 @@ interface userInterface {
 }
 
 const checkEmail = async (email: string) => {
-  try {
-    const response = await instance.post(API_URL + "/check", { email });
-
-    console.log(response);
-
-    return response.data;
-  } catch (error) {
-    return error;
-  }
-};
-
-const getUsers = async () => {
-  try {
-    const response = await instance.get(API_URL + "/");
-
-    console.log(response);
-
-    return response.data;
-  } catch (error) {
-    return error;
-  }
+  const response = await instance.post(API_URL + "/check", { email });
+  return response.data;
 };
 
 // Register user
 const register = async (userData: userInterface) => {
-  try {
-    const response = await instance.post(API_URL + "/register", userData);
-
-    if (response.data) {
-      localStorage.setItem("user", JSON.stringify(response.data));
-    }
-
-    return response.data;
-  } catch (error) {
-    return error;
-  }
-};
-
-// Login user
-const login = async (userData: userInterface) => {
-  const response = await instance.post(API_URL + "/login", userData);
+  const response = await instance.post(API_URL + "/register", userData);
 
   if (response.data) {
-    localStorage.setItem("user", JSON.stringify(response.data));
+    // localStorage.setItem("user", JSON.stringify(response.data));
     console.log(response.data);
   }
 
+  return response.data;
+};
+
+// Login user
+const login = async (userData: userInterface, keepLogIn: boolean) => {
+  const response = await instance.post(API_URL + "/login", userData);
+
+  if (response.data && keepLogIn) {
+    localStorage.setItem("user", JSON.stringify(response.data));
+  }
+  localStorage.removeItem("user");
   return response.data;
 };
 
@@ -75,7 +52,6 @@ const authService = {
   logout,
   login,
   getCurrentUser,
-  getUsers,
 };
 
 export default authService;
